Back end JSON Integration | Data Handling in WPF
Where we continue building our ToDoList application with C# WPF and the MVVM design pattern. This episode is crucial for those interested in learning about data handling and storage in WPF applications.

Creating Methods for JSON Interaction: We dive into how to create methods in C# for reading and writing data to and from a JSON file, a common format for storing application data.
Initialising JSON DataService: Discover how to initialize a JSON data service in your application. This includes checking if a JSON file exists in the data folder and creating one if it doesn't. This is an essential step for robust data management in your app.

Data Persistence in WPF: Understanding how to manage data persistence is key in application development. We cover practical techniques to handle this using JSON.
File Management Techniques: Learn how to check for and manage files within your application, a skill that's widely applicable in software development.
Hands-On Coding Practice: Follow along with real code examples and implement them in your own project for a practical learning experience.